Add much better server status page
This includes automatic status polling based on current status.

The current behavior is to poll when the server is starting up or
shutting down, and then allow manual refreshing when the server is
running or stopped.

This also adds a bit better authentication handling, with a snackbar
popup when an authentication error occurs.